Removed the shrinkwrap Archive as a requirement for using SmallRye OpenAPI.  However, a ShrinkWrap IndexView is still required.
