+++ implementation/src/main/java/io/smallrye/openapi/runtime/OpenApiStaticFile.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2018 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.smallrye.openapi.runtime;
+
+import java.io.Closeable;
+import java.io.IOException;
+import java.io.InputStream;
+
+import io.smallrye.openapi.runtime.io.OpenApiSerializer.Format;
+
+/**
+ * @author eric.wittmann@gmail.com
+ */
+public class OpenApiStaticFile implements Closeable {
+    
+    private Format format;
+    private InputStream content;
+    
+    /**
+     * Constructor.
+     */
+    public OpenApiStaticFile() {
+    }
+
+    /**
+     * Constructor.
+     * @param content
+     * @param format
+     */
+    public OpenApiStaticFile(InputStream content, Format format) {
+        this.setContent(content);
+        this.setFormat(format);
+    }
+
+    /**
+     * @see java.io.Closeable#close()
+     */
+    @Override
+    public void close() throws IOException {
+        if (this.getContent() != null) {
+            this.getContent().close();
+        }
+    }
+
+    /**
+     * @return the format
+     */
+    public Format getFormat() {
+        return format;
+    }
+
+    /**
+     * @param format the format to set
+     */
+    public void setFormat(Format format) {
+        this.format = format;
+    }
+
+    /**
+     * @return the content
+     */
+    public InputStream getContent() {
+        return content;
+    }
+
+    /**
+     * @param content the content to set
+     */
+    public void setContent(InputStream content) {
+        this.content = content;
+    }
+
+}
